[pf-161] refacto of result-item and result-item-campaign
