+++ mon-pix/app/styles/app.scss
@@ -59,7 +59,6 @@
 @import 'components/reset-password-form';
 @import 'components/resume-campaign-banner';
 @import 'components/result-item';
-@import 'components/result-item-campaign';
 @import 'components/results-warning';
 @import 'components/rounded-panel';
 @import 'components/score-pastille';
