--- mon-pix/app/styles/components/_result-item-campaign.scss
@@ -1,116 +1,0 @@
-.result-item-campaign {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  width: auto;
-  height: 110px;
-  border-radius: 4px;
-  border: none;
-  outline: none;
-  padding: 8px 0;
-  background-color: $white;
-
-  &:nth-child(odd) {
-    background: $porcelain-grey;
-  }
-
-  @include device-is('tablet') {
-    height: 70px;
-    padding: 0;
-  }
-}
-
-.result-item-campaign__instruction {
-  width: 82%;
-  font-size: 1.6rem;
-  margin-left: 10px;
-  overflow: hidden;
-
-  @include device-is('tablet') {
-    width: 60%;
-    margin-left: 30px;
-  }
-}
-
-.result-item-campaign__correction {
-  font-size: 1.2rem;
-  margin-right: 26px;
-  margin-left: 35px;
-
-  @include device-is('tablet') {
-    margin-left: auto;
-  }
-}
-
-.result-item-campaign__correction-button {
-  font-weight: $font-normal;
-  font-size: 1.4rem;
-  color: $pix-blue;
-  border: none;
-  outline: none;
-  background-color: transparent;
-
-  &:hover,
-  &:focus,
-  &:active {
-    color: $pix-blue-hover;
-    cursor: pointer;
-    text-decoration: underline;
-  }
-}
-
-.result-item-campaign__icon {
-  margin-left: 10px;
-  font-size: 1.5em;
-
-  @include device-is('tablet') {
-    margin-left: 30px;
-  }
-
-  &--green {
-    color: $mountain-meadow;
-  }
-
-  &--orange {
-    color: $dark-orange;
-  }
-
-  &--red {
-    color: $vermilion;
-  }
-
-  &--grey {
-    color: $dove-gray;
-  }
-
-  /* Forces bootstrap's default by including bootstrap props into .assessment-results container */
-  .tooltip-inner {
-    padding-top: 13px;
-    width: auto;
-    padding-left: 16px;
-    padding-right: 16px;
-    height: 48px;
-    background-color: $mercury-grey;
-    color: $charcoal-grey;
-    font-family: $font-lato;
-    font-size: 1.4rem;
-    box-shadow: 2px 2px 2px 0 rgba(0, 0, 0, 0.2);
-    border: solid 1px $white;
-  }
-
-  .tooltip.top .tooltip-arrow {
-    border-top-color: $mercury-grey;
-  }
-
-  .tooltip.in {
-    opacity: 1;
-    transition: opacity .75s ease;
-  }
-
-  .tooltip.top .tooltip-arrow {
-    bottom: -4px;
-    left: 50%;
-    margin-left: -13px;
-    border-width: 10px 13px 0;
-  }
-}
