--- mon-pix/app/templates/components/result-item-campaign.hbs
@@ -1,15 +1,0 @@
-<div class="result-item-campaign__icon">
-  <div data-toggle="tooltip" data-placement="top" title="{{resultItem.tooltip}}">
-    {{fa-icon resultItem.icon class=(concat 'result-item-campaign__icon--' resultItem.color)}}
-  </div>
-</div>
-
-<div class="result-item-campaign__instruction">
-  {{strip-instruction (convert-to-html answer.challenge.instruction) textLength}}
-</div>
-
-<div class="result-item-campaign__correction">
-  {{#if validationImplementedForChallengeType}}
-    <button {{action openAnswerDetails answer}} class="result-item-campaign__correction-button link js-correct-answer">Réponses et tutos</button>
-  {{/if}}
-</div>
