+++ mon-pix/tests/unit/components/result-item-test.js
@@ -19,14 +19,6 @@ const answerWithNullResult = {
   result: null,
   name: 'answerWithNullResult',
 };
-const answerWithOkResult = {
-  value: '',
-  result: 'ok'
-};
-const answerWithRandomResult = {
-  value: '',
-  result: 'RANDOM_RESULT'
-};
 
 describe('Unit | Component | result-item-component', function() {
 
@@ -45,9 +37,10 @@ describe('Unit | Component | result-item-component', function() {
       answerWithUndefinedResult,
       answerWithNullResult
     ].forEach(function(answer) {
-      it(`should returns false when answer provided is: ${answer.name}`, function() {
+      it(`should returns undefined when answer provided is: ${answer.name}`, function() {
         // when
         component.set('answer', answer);
+
         // then
         expect(component.get('resultItem')).to.be.undefined;
       });
@@ -56,18 +49,88 @@ describe('Unit | Component | result-item-component', function() {
   });
 
   describe('#resultItem Computed property - defined case', function() {
-    it('should returns true when answer provided with result ok', function() {
+    it('should return the green check-circle icon with "Réponse correcte" when answer provided has a valid result', function() {
+      // given
+      const answerWithOkResult = { result: 'ok' };
+
       // when
       component.set('answer', answerWithOkResult);
+
       // then
       expect(component.get('resultItem.tooltip')).to.equal('Réponse correcte');
+      expect(component.get('resultItem.color')).to.equal('green');
+      expect(component.get('resultItem.icon')).to.equal('check-circle');
     });
 
-    it('should returns true when answer provided with result uncommon value by not null or undefined ', function() {
+    it('should return the red times-circle icon with "Réponse incorrecte" when answer provided has an invalid result', function() {
+      // given
+      const answerWithKoResult = { result: 'ko' };
+
       // when
-      component.set('answer', answerWithRandomResult);
+      component.set('answer', answerWithKoResult);
+
       // then
-      expect(component.get('resultItem.tooltip')).to.equal('Correction automatique en cours de développement ;)');
+      expect(component.get('resultItem.tooltip')).to.equal('Réponse incorrecte');
+      expect(component.get('resultItem.color')).to.equal('red');
+      expect(component.get('resultItem.icon')).to.equal('times-circle');
+    });
+
+    it('should return the red times-circle icon with "Temps dépassé" when answer provided has a timed out result', function() {
+      // given
+      const answerWithTimedoutResult = { result: 'timedout' };
+
+      // when
+      component.set('answer', answerWithTimedoutResult);
+
+      // then
+      expect(component.get('resultItem.tooltip')).to.equal('Temps dépassé');
+      expect(component.get('resultItem.color')).to.equal('red');
+      expect(component.get('resultItem.icon')).to.equal('times-circle');
+    });
+
+    it('should return the orange check-circle icon with "Réponse partielle" when answer provided has partially valid result', function() {
+      // given
+      const answerWithPartiallyResult = { result: 'partially' };
+
+      // when
+      component.set('answer', answerWithPartiallyResult);
+
+      // then
+      expect(component.get('resultItem.tooltip')).to.equal('Réponse partielle');
+      expect(component.get('resultItem.color')).to.equal('orange');
+      expect(component.get('resultItem.icon')).to.equal('check-circle');
+    });
+
+    it('should return the grey times-circle icon with "Sans réponse" when answer provided has skipped result', function() {
+      // given
+      const answerWithAbandResult = { result: 'aband' };
+
+      // when
+      component.set('answer', answerWithAbandResult);
+
+      // then
+      expect(component.get('resultItem.tooltip')).to.equal('Sans réponse');
+      expect(component.get('resultItem.color')).to.equal('grey');
+      expect(component.get('resultItem.icon')).to.equal('times-circle');
+    });
+  });
+
+  describe('#textLength', function() {
+
+    it('should be 60 when it is a mobile', function() {
+      //when
+      window.innerWidth = 600;
+
+      //then
+      expect(component.get('textLength')).to.equal(60);
+    });
+
+    it('should be 110 when it a tablet/desktop', function() {
+      //when
+      window.innerWidth = 1200;
+
+      //then
+      expect(component.get('textLength')).to.equal(110);
     });
   });
 
@@ -76,9 +139,10 @@ describe('Unit | Component | result-item-component', function() {
     [
       { challengeType: 'QCM', expected: true },
       { challengeType: 'QROC', expected: true },
-      { challengeType: 'QROCm-ind', expected: false },
-      { challengeType: 'QROCm-dep', expected: false },
-      { challengeType: 'QCU', expected: true }
+      { challengeType: 'QROCM-ind', expected: true },
+      { challengeType: 'QROCM-dep', expected: true },
+      { challengeType: 'QCU', expected: true },
+      { challengeType: 'OtherType', expected: false },
     ].forEach(function(data) {
 
       it(`should return ${data.expected} when challenge type is ${data.challengeType}`, function() {
@@ -93,7 +157,5 @@ describe('Unit | Component | result-item-component', function() {
         expect(component.get('validationImplementedForChallengeType')).to.equal(data.expected);
       });
     });
-
   });
-
 });
